fix: 엔티티 수정

+++ src/main/java/com/gugucon/shopping/item/domain/entity/RateStat.java
@@ -27,7 +27,7 @@ public class RateStat extends BaseTimeEntity {
     @Enumerated(EnumType.STRING)
     private Gender gender;
 
-    private Integer totalScore;
+    private Long totalScore;
 
-    private Integer count;
+    private Long count;
 }
